+++ changelog.md
@@ -7,6 +7,8 @@
 
 ### Mailing
 - Added possibility to hide recipients list.
+- Added possibility to test email host settings without modifying JCM's configuration during the installation process
+ and on Admin>Settings page.
 
 ### Assignments
 - Send reminder to user about his/her upcoming presentation